Extract `dotnet.ts` from utilities folder
